Filled in the data store implementation

+++ src/data_store.h
@@ -24,17 +24,17 @@ struct mdhim_store_cur_opts_t;
 
 /* Function pointers for abstracting data stores */
 typedef int (*mdhim_store_open_fn_t)(void *db_handle, char *path, int flags, struct mdhim_store_opts_t *mstore_opts);
-typedef int (*mdhim_store_put_fn_t)(void *db_handle, void *key, int key_len, void *data, int data_len, 
+typedef int (*mdhim_store_put_fn_t)(void *db_handle, void *key, int key_len, void *data, int64_t data_len, 
 				    struct mdhim_store_opts_t *mstore_opts);
-typedef int (*mdhim_store_bput_fn_t)(void *db_handle, void **keys, int *key_lens, void **data, int *data_lens, 
+typedef int (*mdhim_store_bput_fn_t)(void *db_handle, void **keys, int *key_lens, void **data, int64_t *data_lens, 
 				     struct mdhim_store_opts_t *mstore_opts);
-typedef int (*mdhim_store_get_fn_t)(void *db_handle, void *key, int key_len, void *data, int data_len, 
+typedef int (*mdhim_store_get_fn_t)(void *db_handle, void *key, int key_len, void *data, int64_t data_len, 
 				    struct mdhim_store_opts_t *mstore_opts);
-typedef int (*mdhim_store_bget_fn_t)(void *db_handle, void **keys, int *key_lens, void **data, int *data_lens, 
+typedef int (*mdhim_store_bget_fn_t)(void *db_handle, void **keys, int *key_lens, void **data, int64_t *data_lens, 
 				     struct mdhim_store_opts_t *mstore_opts);
-typedef int (*mdhim_store_get_next_fn_t)(struct mdhim_store_cur_t *mcur, void *data, int *data_len, 
+typedef int (*mdhim_store_get_next_fn_t)(struct mdhim_store_cur_t *mcur, void *data, int64_t *data_len, 
 					 struct mdhim_store_cur_opts_t *mstore_cur_opts);
-typedef int (*mdhim_store_get_prev_fn_t)(struct mdhim_store_cur_t *mcur, void *data, int *data_len, 
+typedef int (*mdhim_store_get_prev_fn_t)(struct mdhim_store_cur_t *mcur, void *data, int64_t *data_len, 
 					 struct mdhim_store_cur_opts_t *mstore_cur_opts);
 typedef int (*mdhim_store_del_fn_t)(void *db_handle, void *key, int key_len,
 				    struct mdhim_store_opts_t *mstore_opts);
